Fix harmless compiler warning introduced by the previous check-in.
FossilOrigin-Name: 06f829e9e57c5aa495f519ad7bb379be611e8e21ef60d36e63d055e45f7a6117

+++ src/vdbemem.c
@@ -327,7 +327,7 @@ void sqlite3VdbeMemZeroTerminateIfAble(Mem *pMem){
   if( pMem->z==0 ) return;
   if( pMem->flags & MEM_Dyn ){
     if( pMem->xDel==sqlite3_free
-     && sqlite3_msize(pMem->z) >= pMem->n+1
+     && sqlite3_msize(pMem->z) >= (u64)(pMem->n+1)
      && pMem->z[pMem->n]==0 
     ){
       pMem->flags |= MEM_Term;
